Add config option for adding LLVM MCA based performance statistics
When optimizing for Cortex-M55 or Cortex-M85, the option `-c llvm_mca`
will run LLVM MCA (Machine Code Analyzer) on the original and optimized
source code and emit the estimated performance characteristics as comments
in the generated source file.

This is useful to get an indicator of the quality of the performance
improvements on M55/M85 even without dedicated hardware.